[Deleted] wrong class deleted. [Rename] `data` package rename to `model`.
