+++ app/src/main/java/kr/swkang/nestedrecyclerview/main/MainActivity.java
@@ -16,9 +16,9 @@ import java.util.List;
 import kr.swkang.nestedrecyclerview.R;
 import kr.swkang.nestedrecyclerview.main.list.MainRvAdapter;
 import kr.swkang.nestedrecyclerview.main.list.MainRvItemDecoration;
-import kr.swkang.nestedrecyclerview.main.list.data.Contents;
-import kr.swkang.nestedrecyclerview.main.list.data.subcontents.BodySection;
-import kr.swkang.nestedrecyclerview.main.list.data.subcontents.HeaderContents;
+import kr.swkang.nestedrecyclerview.main.list.model.Contents;
+import kr.swkang.nestedrecyclerview.main.list.model.subcontents.BodySection;
+import kr.swkang.nestedrecyclerview.main.list.model.subcontents.HeaderContents;
 import kr.swkang.nestedrecyclerview.utils.BaseActivity;
 import kr.swkang.nestedrecyclerview.utils.OnViewClickListener;
 import kr.swkang.nestedrecyclerview.utils.rvs.SwRecyclerView;
