+++ app/src/main/java/kr/swkang/nestedrecyclerview/main/list/MainRvAdapter.java
@@ -3,8 +3,6 @@ package kr.swkang.nestedrecyclerview.main.list;
 import android.content.Context;
 import android.support.annotation.NonNull;
 import android.support.v4.app.FragmentManager;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.text.TextUtils;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -19,12 +17,12 @@ import java.util.ArrayList;
 
 import kr.swkang.nestedrecyclerview.R;
 import kr.swkang.nestedrecyclerview.main.header.HeaderViewPagerAdapter;
-import kr.swkang.nestedrecyclerview.main.list.data.Contents;
-import kr.swkang.nestedrecyclerview.main.list.data.ContentsType;
-import kr.swkang.nestedrecyclerview.main.list.data.SectionHeader;
-import kr.swkang.nestedrecyclerview.main.list.data.subcontents.BodyItems;
-import kr.swkang.nestedrecyclerview.main.list.data.subcontents.BodySection;
-import kr.swkang.nestedrecyclerview.main.list.data.subcontents.HeaderContents;
+import kr.swkang.nestedrecyclerview.main.list.model.Contents;
+import kr.swkang.nestedrecyclerview.main.list.model.ContentsType;
+import kr.swkang.nestedrecyclerview.main.list.model.SectionHeader;
+import kr.swkang.nestedrecyclerview.main.list.model.subcontents.BodyItems;
+import kr.swkang.nestedrecyclerview.main.list.model.subcontents.BodySection;
+import kr.swkang.nestedrecyclerview.main.list.model.subcontents.HeaderContents;
 import kr.swkang.nestedrecyclerview.utils.OnViewClickListener;
 import kr.swkang.nestedrecyclerview.utils.rvs.SwRecyclerViewAdapter;
 import kr.swkang.nestedrecyclerview.utils.rvs.snappy.SnapRecyclerView;
