--- app/src/main/java/kr/swkang/nestedrecyclerview/utils/mvp/BaseModel.java
@@ -1,9 +1,0 @@
-package kr.swkang.nestedrecyclerview.utils.mvp;
-
-/**
- * @author KangSung-Woo
- * @since 2016/05/19
- */
-public class BaseModel {
-
-}
